Refactor server.js middleware configurations & fix ESM test mocks
